yeh, the code is actually easy to read now (that i have more time)
* and i guess (:load-after X) after (setup (:pkg y) ... ) doesn't make much
sense ... but it really depends on how emacs implements a reactor pattern to
delegate compute to single thread ... and how this changes once code is
byte/compiled... and whether guix or straight are loading deps ... and
probably other things.

not everyone who's career was assassinated is as stupied as they were racked up
to be. the problem of course is that such persists for only as energy/burnrate
is available to maintain the imbalance. most people depend on socially mediated
benefits. no, i didn't benefit from that for a DECADE. but the conditions for
victory are such: you contain me /forever/, you destroy one life. i get out /one
good time/, i have a normal life. what cannot go on forever will surely come to
an end.

"first they ignore you, then they call you crazy..."